# src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: E-Commerce Mock API
  version: 1.0.0
  description: A mock API for an e-commerce platform
servers:
  - url: http://localhost:8080
paths:
  /api/user/{userId}/balance:
    get:
      summary: 사용자 잔액 조회
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 잔액 조회 성공
          content:
            application/json:
              schema:
                type: number
                format: double
  /api/user/balance/charge:
    post:
      summary: 포인트 충전
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                points:
                  type: number
                  format: double
              required:
                - userId
                - points
      responses:
        '200':
          description: 포인트 충전 성공
          content:
            application/json:
              schema:
                type: string
  /api/products:
    get:
      summary: 상품 조회
      responses:
        '200':
          description: 상품 목록 조회 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    productId:
                      type: integer
                    name:
                      type: string
                    price:
                      type: number
                      format: double
                    stockQuantity:
                      type: integer
  /api/order/place:
    post:
      summary: 상품 주문
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      productId:
                        type: string
                      quantity:
                        type: integer
                      price:
                        type: number
                        format: double
              required:
                - userId
                - items
      responses:
        '200':
          description: 주문 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  success:
                    type: boolean
        '400':
          description: 잔액 부족
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  success:
                    type: boolean
  /api/cart/add:
    post:
      summary: 장바구니 추가
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                productId:
                  type: string
                quantity:
                  type: integer
              required:
                - userId
                - productId
                - quantity
      responses:
        '200':
          description: 장바구니에 상품 추가 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  success:
                    type: boolean
  /api/cart/remove:
    patch:
      summary: 장바구니에서 상품 삭제
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                productId:
                  type: string
              required:
                - userId
                - productId
      responses:
        '200':
          description: 장바구니에서 상품 삭제 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  success:
                    type: boolean